+++ app/calendar/page.tsx
@@ -1,17 +1,11 @@
 "use client";
 
 import React, { Suspense, useCallback, useMemo, useState } from "react";
-import { useQuery } from "@tanstack/react-query";
-import { ExhibitionCardList } from "@/components/pages/Home/ExhibitionCardList/ExhibitionCardList";
-import { Select } from "@/components/ui/Select/Select";
-import { useSelectCategory } from "@/hooks/useSelectCategory";
-import { AppBar } from "@/components/pages/Home/AppBar/AppBar";
-import { Exhibition } from "@/interfaces/exhibition";
-import { getExhibitionList, togglePinById } from "@/apis/exhibition";
-import { getCategories } from "@/apis/category";
-import { PostFloatingButton } from "@/components/pages/Home/PostFloatingButton/PostFloatingButton";
-import { sendMessage } from "@/libs/message/message";
 import * as S from "./page.styles";
+import NavigationBar from "@/components/ui/NavigationBar/NavigationBar";
+import Calendar from "@/components/pages/Calendar/Calendar/Calendar";
+import "antd-mobile/bundle/style.css";
+import { toYYYYMMDD } from "@/utils/datetime";
 
 export default function PageWrapper() {
   return (
@@ -22,75 +16,20 @@ export default function PageWrapper() {
 }
 
 function Page() {
-  const categoriesQuery = useQuery({
-    queryKey: ["categories"],
-    queryFn: getCategories,
-  });
-
-  const { selectedIndex, selectCategoryByIndex: handleSelectCategory } =
-    useSelectCategory();
-
-  const {
-    selectedIndex: selectedFilter,
-    selectCategoryByIndex: handleSelectFilter,
-  } = useSelectCategory();
-
-  const exhibitionListQuery = useQuery({
-    queryKey: ["exhibitionList"],
-    queryFn: getExhibitionList,
-  });
-
-  const [pins, setPins] = useState<Record<string, boolean>>({});
-  const exhibitionListWithPin = useMemo(() => {
-    return (exhibitionListQuery.data ?? []).map((item) => {
-      return {
-        ...item,
-        isPin: Boolean(pins[item.id]),
-      };
-    });
-  }, [exhibitionListQuery.data, pins]);
-  const [fixedExhibition, ...restExhibition] = exhibitionListWithPin;
-
-  const handleRegisterCategory = useCallback(() => {}, []);
-
-  const handleTogglePin = useCallback(
-    async (e: React.MouseEvent, item: Exhibition) => {
-      const id = String(item.id);
-      setPins((pins) => {
-        return {
-          ...pins,
-          [id]: !Boolean(pins[id]),
-        };
-      });
-      await togglePinById(id);
-    },
-    []
-  );
-
-  const handleEditButton = useCallback((e: React.MouseEvent) => {
-    sendMessage(["NAVIGATE_TO_EDIT"]);
-  }, []);
+  const bgImages = {
+    [toYYYYMMDD()]: "https://picsum.photos/358"
+  }
+  const handleGoBackClick = () => {};
 
   return (
     <S.Wrapper>
-      <AppBar />
-      <S.CategoryListStyled
-        activeIndex={selectedIndex}
-        items={categoriesQuery?.data ?? []}
-        onSelected={handleSelectCategory}
-        onRegister={handleRegisterCategory}
+      <NavigationBar
+        goBack={{ name: "ArrowLeftIcon", size: 24 }}
+        onGoBackClick={handleGoBackClick}
       />
-      <S.Filter>
-        <Select activeIndex={selectedFilter} onSelected={handleSelectFilter} />
-      </S.Filter>
       <S.Content>
-        <ExhibitionCardList
-          fixedExhibition={fixedExhibition}
-          exhibitionList={restExhibition}
-          onTogglePin={handleTogglePin}
-        />
+        <Calendar bgImages={bgImages} />
       </S.Content>
-      <PostFloatingButton onClick={handleEditButton} />
     </S.Wrapper>
   );
 }
