--- components/ui/Calendar/Calendar/Calendar.stories.tsx
@@ -1,13 +1,0 @@
-import { Story, Meta } from '@storybook/react';
-import { Calendar, Props } from './Calendar';
-
-export default {
-  component: Calendar,
-  title: 'UI/Calendar/Calendar',
-  argTypes: {},
-} as Meta;
-
-const Template: Story<Props> = (args) => <Calendar {...args} />;
-
-export const Primary = Template.bind({});
-Primary.args = {};
