+++ libs/antd-mobile-provider.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import React from "react";
+import { ConfigProvider } from "antd-mobile";
+import koKR from "antd-mobile/es/locales/ko-KR";
+
+export default function AntdMobileProvider({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <ConfigProvider locale={koKR}>{children}</ConfigProvider>;
+}
