+++ utils/datetime.ts
@@ -1,0 +1,3 @@
+export const toYYYYMMDD = (date: Date = new Date()) => {
+  return `${date.getFullYear()}${date.getMonth() + 1}${date.getDate()}}`;
+};
